[issue-456] - workaround for issue-488

+++ Mini-Lisp.Chic/hacks.h
@@ -23,9 +23,9 @@
 #define let static
 typedef const char *const String;
 
-typedef int32_t W; // A full word of  32 bits; two half words
-typedef int16_t H; // Half a word including 16 bits.
-typedef int8_t byte; // A byte 
+typedef int64_t W; // A full word of  64 bits; two half words
+typedef int32_t H; // Half a word including 32 bits.
+typedef int8_t byte; // A byte
 
 static inline H size(String s) { for (H $ = 0;; ++$) if (s[$] == '\0') return $ + 1; }
 #endif
